+++ src/api/registration/CodeTransform.ts
@@ -19,6 +19,8 @@ import { Project } from "@atomist/automation-client/project/Project";
 import { NoParameters } from "@atomist/automation-client/SmartParameters";
 import { ParametersInvocation } from "../listener/ParametersInvocation";
 
+export type TransformResult = EditResult;
+
 /**
  * Function that can transform a project. Mixing HandlerContextMethods into second
  * parameter, and third parameter are only for backward compatibility.
@@ -26,7 +28,7 @@ import { ParametersInvocation } from "../listener/ParametersInvocation";
  */
 export type CodeTransform<P = NoParameters> = (p: Project,
                                                sdmc: ParametersInvocation<P>,
-                                               params?: P) => Promise<Project | EditResult>;
+                                               params?: P) => Promise<Project | TransformResult>;
 
 /**
  * One or many CodeTransforms