+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 -   Push tests for adding a goal implementation are not run. [#490](https://github.com/atomist/sdm/issues/490)
 -   Transform-testNaming. [#502](https://github.com/atomist/sdm/issues/502)
 -   **BREAKING** Remove unimplemented property observesOnly. [#501](https://github.com/atomist/sdm/issues/501)
+-   **BREAKING** No longer export `EditResult`. Use `TransformResult`
 
 ### Fixed